# src/main/java/com/example/demo/model/User.java
package com.example.demo.model;

import jakarta.persistence.*;

import java.io.Serializable;

@Entity
@Table(name = "users")
public class User implements Serializable {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(name = "login")
    private String login;

    private String firstName;
    private String lastName;

    @Column(name = "email")
    private String email;
    private boolean activated;

    @Column(name = "lang_key")
    private String langKey;

    public User() {}

    public User(String login, String firstName, String lastName, String email, boolean activated, String langKey) {
        this.login = login;
        this.firstName = firstName;
        this.lastName = lastName;
        this.email = email;
        this.activated = activated;
        this.langKey = langKey;
    }

    public User(AdminUserDTO adminUserDTO) {
        this.login = adminUserDTO.getLogin();
        this.firstName = adminUserDTO.getFirstName();
        this.lastName = adminUserDTO.getLastName();
        this.email = adminUserDTO.getEmail();
        this.activated = adminUserDTO.isActivated();
        this.langKey = adminUserDTO.getLangKey();
    }

    // Getters y Setters
    public String getLogin() {
        return login;
    }

    public void setLogin(String login) {
        this.login = login;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public boolean isActivated() {
        return activated;
    }

    public void setActivated(boolean activated) {
        this.activated = activated;
    }

    public String getLangKey() {
        return langKey;
    }

    public void setLangKey(String langKey) {
        this.langKey = langKey;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getId() {
        return id;
    }
}